requirements.txt
diffusers==0.25.0
av
onnxruntime
peft
pytorch_lightning
ruamel.yaml<0.18.0
mediapipe
ffmpeg

omegaconf==2.3.0
jaxtyping
typeguard
nerfacc==0.3.3
pyrender
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
accelerate
opencv-python
tensorboard
matplotlib
imageio>=2.28.0
imageio[ffmpeg]
git+https://github.com/NVlabs/nvdiffrast.git
libigl
xatlas
trimesh[easy]
networkx
pysdf
PyMCubes
wandb
gradio
git+https://github.com/ashawkey/envlight.git
torchmetrics

# deepfloyd
bitsandbytes==0.38.1
sentencepiece
safetensors
huggingface_hub==0.25.0

# for zero123
einops
kornia==0.7.0
taming-transformers-rom1504
git+https://github.com/openai/CLIP.git

#controlnet
controlnet_aux

#vis
torch_efficient_distloss
mediapy
plyfile
rembg
git+https://github.com/ashawkey/kiuikit
scikit-learn
scipy
viser
ninja
easydict
albumentations==0.5.2
webdataset


#wonder3d
nerfacc==0.3.3
pyrender==0.1.45
Pillow==9.5.0


# gaussian splatting
stage2/gaussiansplatting/submodules/diff-gaussian-rasterization
stage2/gaussiansplatting/submodules/simple-knn